move stub code into stubresolver.cc/hh
