+++ pdns/secpoll-auth.hh
@@ -2,9 +2,9 @@
 #define PDNS_SECPOLL_AUTH_HH
 #include <time.h>
 #include "namespaces.hh"
+#include "stubresolver.hh"
 
 void doSecPoll(bool first);
-void secPollParseResolveConf();
 extern std::string g_security_message;
 
 #endif
